demo/python/requirements.txt
opentelemetry-api>=1.28.1,<2.0.0
requests>=2.31.0,<3.0.0
opentelemetry-instrumentation-requests>=0.50b0,<1.0.0
opentelemetry-semantic-conventions>=0.50b0,<1.0.0
otlp-stdout-adapter>=0.3.0,<1.0.0
lambda-otel-lite>=0.5,<1.0.0

